Update Callback page design, remove /login route, refactor main component in App.tsx
