+++ src/components/login/CallbackPage.tsx
@@ -4,11 +4,17 @@ import { loginWithPkce } from '../../store/account/action';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { deepEqual } from '../../utils/common';
+import { Home } from '@mui/icons-material';
+import HomeElement from '../home/HomeElement';
+import { useSelector } from 'react-redux';
+import { AppState } from '../../store';
 
 const CallbackPage: React.FC = () => {
   const [tokenResponse, setTokenResponse] = useState<any>({})
+  const { authenticated } = useSelector((state: AppState) => state.account);
   const dispatch = useDispatch();
   const navigate = useNavigate()
+  const [displayText, setDisplayText] = useState(authenticated ? "Already successfully authenticated." : "Waiting to be authenticated...")
   
   useEffect(() => {
     const searchParams = new URLSearchParams(window.location.search);
@@ -18,11 +24,13 @@ const CallbackPage: React.FC = () => {
       async function getTokenResponse() {
         const newToken = await handleCallback(oidcConfigUrl, clientId, redirectUri)
         if (newToken.error) {
-          console.error('Error fetching token:', newToken.error)
+          console.error('Fetched token error:', newToken.error)
+          setDisplayText("Error fetching token. Please try again.")
         } else {
           localStorage.setItem('user_token', JSON.stringify(newToken))
           await dispatch(loginWithPkce(newToken) as any)
           setTokenResponse(newToken)
+          setDisplayText("Successfully authenticated! You can now access Admin UI.")
         }
       }
       const oidcConfigUrl = sessionStorage.getItem('oidc_config_url');
@@ -30,7 +38,8 @@ const CallbackPage: React.FC = () => {
       const redirectUri = sessionStorage.getItem('redirect_uri');
       getTokenResponse()
     } else if (error) {
-      console.error('Error fetching token:', error)
+      setDisplayText("Error authenticating. Please try again.")
+      console.error('Failed to initialize authorization request:', error)
     }
   }, [dispatch, navigate])
 
@@ -41,11 +50,19 @@ const CallbackPage: React.FC = () => {
     }
   }, [tokenResponse, navigate])
 
+  const CallbackElement: React.FC = () => {
+    return (
+      <section style={{ display: 'flex', flexDirection: 'column', padding: '40px 0', gap: '10px' }}>
+        <span style={{ float: 'left', fontSize: '24px', color: 'black'}}>HCL Domino REST API Administrator</span>
+        <h1>{displayText}</h1>
+      </section>
+    )
+  }
+
   return (
-    <div>
-      <h1>Hello</h1>
-      <p>{`${JSON.stringify(tokenResponse)}`}</p>
-    </div>
+    <>
+      <HomeElement MainElement={CallbackElement} />
+    </>
   );
 };
