+++ src/components/routers/ProtectedRoute.tsx
@@ -5,6 +5,6 @@ import { AppState } from '../../store';
 export const PrivateRoutes = () => {
     const { authenticated } = useSelector((state: AppState) => state.account);
     return (
-        authenticated ? <Outlet/> : <Navigate to='/login'/>
+        authenticated ? <Outlet/> : <Navigate to='/'/>
     )
 }
